Changed the type of some variables to fixed size integers

+++ include/cpu.h
@@ -20,7 +20,7 @@ namespace gameboy
     namespace cpu_cycles
     {
         // clang-format off
-        constexpr int OPCODE_CYCLES[256] =
+        constexpr uint8_t OPCODE_CYCLES[256] =
         {// 0  1  2  3  4  5  6  7  8  9  A  B  C  D  E  F
             1, 3, 2, 2, 1, 1, 2, 1, 5, 2, 2, 2, 1, 1, 2, 1, // 0
             1, 3, 2, 2, 1, 1, 2, 1, 3, 2, 2, 2, 1, 1, 2, 1, // 1
@@ -40,7 +40,7 @@ namespace gameboy
             3, 3, 2, 1, 0, 4, 2, 4, 3, 2, 4, 1, 0, 0, 2, 4  // F
         }; ///< Opcodes machine cycles
 
-        constexpr int OPCODE_CYCLES_BRANCHED[256] =
+        constexpr uint8_t OPCODE_CYCLES_BRANCHED[256] =
         {// 0  1  2  3  4  5  6  7  8  9  A  B  C  D  E  F
             1, 3, 2, 2, 1, 1, 2, 1, 5, 2, 2, 2, 1, 1, 2, 1, // 0
             1, 3, 2, 2, 1, 1, 2, 1, 3, 2, 2, 2, 1, 1, 2, 1, // 1
@@ -60,7 +60,7 @@ namespace gameboy
             3, 3, 2, 1, 0, 4, 2, 4, 3, 2, 4, 1, 0, 0, 2, 4  // F
         }; ///< Opcodes machine cycles when branched (jump, call, return)
 
-        constexpr int OPCODE_CB_CYCLES[256] =
+        constexpr uint8_t OPCODE_CB_CYCLES[256] =
         {// 0  1  2  3  4  5  6  7  8  9  A  B  C  D  E  F
             2, 2, 2, 2, 2, 2, 4, 2, 2, 2, 2, 2, 2, 2, 4, 2, // 0
             2, 2, 2, 2, 2, 2, 4, 2, 2, 2, 2, 2, 2, 2, 4, 2, // 1
@@ -100,7 +100,7 @@ namespace gameboy
          *
          * @return The number of cycles used by the instruction.
          */
-        int cycle();
+        uint8_t cycle();
 
     private:
         Memory &m_memory; ///< The memory
@@ -129,7 +129,7 @@ namespace gameboy
          * @return 0 if no interrupt has been executed, 5 if an interrupt has been executed.
          * @see handleInterrupt
          */
-        int handleInterrupts();
+        uint8_t handleInterrupts();
 
         /**
          * @brief Execute the interrupt.
@@ -148,7 +148,7 @@ namespace gameboy
          * @param opcode The opcode of the instruction.
          * @return The number of cycles used by the instruction.
          */
-        int executeOpcode(uint8_t opcode);
+        uint8_t executeOpcode(uint8_t opcode);
 
         /**
          * @brief Executes the next instruction.
@@ -157,7 +157,7 @@ namespace gameboy
          * @param opcode The cb opcode of the instruction.
          * @return The number of cycles used by the instruction.
          */
-        int executeOpcodeCB(uint8_t opcode);
+        uint8_t executeOpcodeCB(uint8_t opcode);
 
         // Push and pop
         /**
